Split lora.py to lora.py and layers.py
