+++ python/sglang/srt/lora/backend/flashinfer_backend.py
@@ -3,7 +3,7 @@ from typing import Tuple
 import torch
 
 from sglang.srt.lora.backend import BaseLoRABackend
-from sglang.srt.lora.lora import LoRABatchInfo
+from sglang.srt.lora.utils import LoRABatchInfo
 from sglang.srt.utils import is_flashinfer_available
 
 if is_flashinfer_available():
